# Optimizations for ObjectExtensions