- Make MemberResult a struct
- Add a custom key type for the cache to avoid building a string every time
- Use the custom key to prevent closure allocation in the ConcurrentDictionary.GetOrAdd callbacks
- Make the PropertyFetcher generic to prevent boxing when fetching a value type
- Add a Fetch overload to provide the type when it's already known


**Benchmark:**
```csharp
return ObjectsExtensions.Emit.ObjectExtensions.GetProperty<int>("source", "Length").Value;
```

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19041.329 (2004/?/20H1)
Intel Core i7-9750H CPU 2.60GHz, 1 CPU, 12 logical and 6 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4180.0), X64 RyuJIT
  DefaultJob : .NET Framework 4.8 (4.8.4180.0), X64 RyuJIT


```
|          Method |      Mean |    Error |   StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|---------------- |----------:|---------:|---------:|------:|-------:|------:|------:|----------:|
| OldGetProperty | 445.22 ns | 3.179 ns | 2.818 ns |  1.00 | 0.1721 |     - |     - |    1083 B |
| NewGetProperty |  84.56 ns | 0.821 ns | 0.768 ns |  0.19 |      - |     - |     - |         - |