@kevingosse 2020-06-24
Pushed a new version that uses expressions to fetch the property value.

I thought I was being clever with the `where TProperty : T` generic constraint in the property fetcher, but I didn't consider that there isn't necessarily an inheritance relationship between `TProperty` and `T`. For instance, in the HttpMessageHandler integration, we fetch a HttpStatusCode (of type Enum) as an int. Thankfully, this was caught in the integration tests. 

With the latest commit:
- I now generate an expression to fetch the property value. It's probably slower the first time, but it's actually slightly faster when invoked
- With that change, I saw no value in keeping the `TypedFetchProperty` implementation, so I collapsed it into the parent `PropertyFetch`
- I added a unit test for the "enum to int" case

New benchmark:

``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.19041.329 (2004/?/20H1)
Intel Core i7-9750H CPU 2.60GHz, 1 CPU, 12 logical and 6 physical cores
  [Host]     : .NET Framework 4.8 (4.8.4180.0), X64 RyuJIT
  DefaultJob : .NET Framework 4.8 (4.8.4180.0), X64 RyuJIT


```
|                 Method |      Mean |    Error |   StdDev | Ratio |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|----------------------- |----------:|---------:|---------:|------:|-------:|------:|------:|----------:|
|        OldGetProperty  | 441.11 ns | 2.361 ns | 2.093 ns |  1.00 | 0.1721 |     - |     - |    1083 B |
|        NewGetProperty  |  83.61 ns | 0.311 ns | 0.259 ns |  0.19 |      - |     - |     - |         - |
| ExpressionGetProperty  |  82.77 ns | 0.297 ns | 0.232 ns |  0.19 |      - |     - |     - |         - |
|  CollapsedGetProperty  |  76.10 ns | 0.334 ns | 0.296 ns |  0.17 |      - |     - |     - |         - |

NewGetProperty => Previous version of the code (that didn't work with the enum -> int conversion
ExpressionGetProperty => Version that uses expressions in the TypedFetchProperty
CollapsedGetProperty => Version that collapses the TypedFetchProperty into the PropertyFetch (the one I commited)